+++ assets/js/customize-preview.js
@@ -5,7 +5,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0
- * @version  2.2.0
+ * @version  2.5.0
  */
 
 
@@ -13,6 +13,8 @@
 
 
 ( function( $ ) {
+
+	'use strict';
 
 
 
@@ -29,7 +31,7 @@
 				value
 					.bind( function( to ) {
 
-						$( '.site-title span' )
+						$( '.site-title' )
 							.text( to );
 
 					} );
@@ -49,6 +51,41 @@
 					} );
 
 		} ); // /blogdescription
+
+		wp.customize( 'header_textcolor', function( value ) {
+
+			// Processing
+
+				value
+					.bind( function( to ) {
+
+						if ( 'blank' === to ) {
+
+							$( '.site-title, .site-description' )
+								.css( {
+									'clip'     : 'rect(1px, 1px, 1px, 1px)',
+									'position' : 'absolute',
+								} );
+
+							$( 'body' )
+								.addClass( 'site-title-hidden' );
+
+						} else {
+
+							$( '.site-title, .site-description' )
+								.css( {
+									'clip'     : 'auto',
+									'position' : 'relative',
+								} );
+
+							$( 'body' )
+								.removeClass( 'site-title-hidden' );
+
+						}
+
+					} );
+
+		} ); // /header_textcolor