+++ includes/setup/setup.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0
- * @version  2.4.0
+ * @version  2.5.0
  *
  * Contents:
  *
@@ -34,6 +34,10 @@
 	// Custom header
 
 		require_once( get_template_directory() . '/includes/custom-header/custom-header.php' );
+
+	// SVG
+
+		require_once( get_template_directory() . '/includes/svg/class-svg.php' );
 
 
 
@@ -263,7 +267,7 @@
 	 * Theme setup
 	 *
 	 * @since    1.0
-	 * @version  2.2.0
+	 * @version  2.5.0
 	 */
 	if ( ! function_exists( 'wm_setup' ) ) {
 		function wm_setup() {
@@ -278,7 +282,7 @@
 
 					$visual_editor_css = array_filter( (array) apply_filters( 'wmhook_wm_setup_visual_editor_css', array(
 							str_replace( ',', '%2C', wm_google_fonts_url() ),
-							esc_url_raw( add_query_arg( array( 'ver' => wp_get_theme( get_template() )->get( 'Version' ) ), wm_get_stylesheet_directory_uri( 'assets/fonts/genericons/genericons.css' ) ) ),
+							esc_url_raw( add_query_arg( array( 'ver' => wp_get_theme( get_template() )->get( 'Version' ) ), wm_get_stylesheet_directory_uri( 'assets/fonts/genericons-neue/genericons-neue.css' ) ) ),
 							esc_url_raw( add_query_arg( array( 'ver' => wp_get_theme( get_template() )->get( 'Version' ) ), wm_get_stylesheet_directory_uri( 'assets/css/editor-style' . $rtl_suffix . '.css' ) ) ),
 						) ) );
 
@@ -366,12 +370,13 @@
 					 * @link  https://codex.wordpress.org/Function_Reference/add_theme_support#Custom_Header
 					 */
 					add_theme_support( 'custom-header', apply_filters( 'wmhook_wm_setup_custom_header_args', array(
-							'random-default' => true,
-							'header-text'    => false,
-							'width'          => ( isset( $image_sizes['auberge_banner'] ) ) ? ( $image_sizes['auberge_banner'][0] ) : ( 1920 ),
-							'height'         => ( isset( $image_sizes['auberge_banner'] ) ) ? ( $image_sizes['auberge_banner'][1] ) : ( 1080 ),
-							'flex-height'    => true,
-							'flex-width'     => true,
+							'default-text-color' => 'ffffff',
+							'header-text'        => true,
+							'random-default'     => true,
+							'width'              => ( isset( $image_sizes['auberge_banner'] ) ) ? ( $image_sizes['auberge_banner'][0] ) : ( 1920 ),
+							'height'             => ( isset( $image_sizes['auberge_banner'] ) ) ? ( $image_sizes['auberge_banner'][1] ) : ( 1080 ),
+							'flex-height'        => true,
+							'flex-width'         => true,
 						) ) );
 
 					// Default custom headers packed with the theme (thumbnail size: 275x155 px)
@@ -1110,7 +1115,7 @@
 	 * Registering theme styles and scripts
 	 *
 	 * @since    1.0
-	 * @version  2.2.0
+	 * @version  2.5.0
 	 */
 	if ( ! function_exists( 'wm_register_assets' ) ) {
 		function wm_register_assets() {
@@ -1125,7 +1130,7 @@
 				// Styles
 
 					$register_styles = apply_filters( 'wmhook_wm_register_assets_register_styles', array(
-							'genericons'      => array( wm_get_stylesheet_directory_uri( 'assets/fonts/genericons/genericons.css' ) ),
+							'genericons-neue' => array( wm_get_stylesheet_directory_uri( 'assets/fonts/genericons-neue/genericons-neue.css' ) ),
 							'slick'           => array( wm_get_stylesheet_directory_uri( 'assets/css/slick.css' ) ),
 							'wm-google-fonts' => array( wm_google_fonts_url() ),
 							'wm-main'         => array( wm_get_stylesheet_directory_uri( 'assets/css/main.css' ) ),
@@ -1171,7 +1176,7 @@
 	 * Frontend HTML head assets enqueue
 	 *
 	 * @since    1.0
-	 * @version  2.2.0
+	 * @version  2.5.0
 	 */
 	if ( ! function_exists( 'wm_enqueue_assets' ) ) {
 		function wm_enqueue_assets() {
@@ -1216,7 +1221,7 @@
 
 					// Main
 
-						$enqueue_styles[] = 'genericons';
+						$enqueue_styles[] = 'genericons-neue';
 						$enqueue_styles[] = 'wm-main';
 						$enqueue_styles[] = 'wm-stylesheet';
 
@@ -1377,7 +1382,7 @@
 	 * HTML Body classes
 	 *
 	 * @since    1.0
-	 * @version  2.4.0
+	 * @version  2.5.0
 	 *
 	 * @param  array $classes
 	 */
@@ -1401,6 +1406,12 @@
 						$body_classes['not-front-page'] = ++$i;
 					}
 
+				// Is site branding text displayed?
+
+					if ( 'blank' === get_header_textcolor() ) {
+						$body_classes['site-title-hidden'] = ++$i;
+					}
+
 				// Singular?
 
 					if ( is_singular() ) {
@@ -1410,7 +1421,7 @@
 				// Privacy Policy page
 
 					if ( (int) get_option( 'wp_page_for_privacy_policy' ) === get_the_ID() ) {
-						$classes[] = 'page-privacy-policy';
+						$body_classes['page-privacy-policy'] = ++$i;
 					}
 
 				// Has featured image?
@@ -1703,7 +1714,7 @@
 
 
 		/**
-		 * Display social links
+		 * Display social links.
 		 *
 		 * @since    1.0
 		 * @version  2.0
@@ -1719,6 +1730,197 @@
 		} // /wm_menu_social
 
 		add_action( 'tha_header_top', 'wm_menu_social', 130 );
+
+
+
+			/**
+			 * Flush social menu cache
+			 *
+			 * @since    2.5.0
+			 * @version  2.5.0
+			 */
+			function wm_social_cache_flush() {
+
+				// Processing
+
+					delete_transient( 'auberge_social_links' );
+
+			} // /wm_social_cache_flush
+
+			add_action( 'wp_update_nav_menu', 'wm_social_cache_flush' );
+			add_action( 'customize_save_after', 'wm_social_cache_flush' );
+			add_action( 'wmhook_theme_upgrade', 'wm_social_cache_flush' );
+
+
+
+			/**
+			 * Social links supported icons.
+			 *
+			 * @since    2.5.0
+			 * @version  2.5.0
+			 */
+			function wm_social_links_icons() {
+
+				// Output
+
+					return array(
+						'behance.net'       => 'behance',
+						'bitbucket.org'     => 'bitbucket',
+						'codepen.io'        => 'codepen',
+						'deviantart.com'    => 'deviantart',
+						'digg.com'          => 'digg',
+						'docker.com'        => 'dockerhub',
+						'dribbble.com'      => 'dribbble',
+						'dropbox.com'       => 'dropbox',
+						'facebook.com'      => 'facebook',
+						'flickr.com'        => 'flickr',
+						'foursquare.com'    => 'foursquare',
+						'plus.google.com'   => 'google-plus',
+						'github.com'        => 'github',
+						'instagram.com'     => 'instagram',
+						'linkedin.com'      => 'linkedin',
+						'mailto:'           => 'envelope',
+						'medium.com'        => 'medium',
+						'paypal.com'        => 'paypal',
+						'pscp.tv'           => 'periscope',
+						'pinterest.com'     => 'pinterest',
+						'getpocket.com'     => 'get-pocket',
+						'reddit.com'        => 'reddit',
+						'/feed'             => 'rss',
+						'skype.com'         => 'skype',
+						'skype:'            => 'skype',
+						'slack.com'         => 'slack',
+						'slideshare.net'    => 'slideshare',
+						'snapchat.com'      => 'snapchat',
+						'soundcloud.com'    => 'soundcloud',
+						'spotify.com'       => 'spotify',
+						'stackoverflow.com' => 'stack-overflow',
+						'stumbleupon.com'   => 'stumbleupon',
+						'trello.com'        => 'trello',
+						'tripadvisor.'      => 'tripadvisor',
+						'tumblr.com'        => 'tumblr',
+						'twitch.tv'         => 'twitch',
+						'twitter.com'       => 'twitter',
+						'vimeo.com'         => 'vimeo',
+						'vine.co'           => 'vine',
+						'vk.com'            => 'vk',
+						'wordpress.org'     => 'wordpress',
+						'wordpress.com'     => 'wordpress',
+						'xing.com'          => 'xing',
+						'yelp.com'          => 'yelp',
+						'youtube.com'       => 'youtube',
+					);
+
+			} // /wm_social_links_icons
+
+			add_filter( 'wmhook_auberge_svg_get_social_icons', 'wm_social_links_icons' );
+
+
+
+			/**
+			 * Display SVG icons in social links menu.
+			 *
+			 * @since    2.5.0
+			 * @version  2.5.0
+			 *
+			 * @param  string  $item_output The menu item output.
+			 * @param  WP_Post $item        Menu item object.
+			 * @param  int     $depth       Depth of the menu.
+			 * @param  array   $args        wp_nav_menu() arguments.
+			 */
+			function wm_nav_menu_social_icons( $item_output, $item, $depth, $args ) {
+
+				// Variables
+
+					$locations = get_nav_menu_locations();
+
+
+				// Processing
+
+					if (
+						isset( $locations['social'] )
+						&& isset( $args->menu->term_id )
+						&& absint( $locations['social'] ) === absint( $args->menu->term_id )
+					) {
+
+						$social_icons = Auberge_SVG::get_social_icons();
+						$social_icon  = 'chain';
+
+						foreach ( $social_icons as $url => $icon ) {
+							if ( false !== strpos( $item_output, $url ) ) {
+								$social_icon = $icon;
+								break;
+							}
+						}
+
+						$item_output = str_replace(
+							$args->link_after,
+							'</span>' . Auberge_SVG::get( array(
+								'icon' => esc_attr( $social_icon ),
+								'base' => 'social-icon',
+							) ),
+							$item_output
+						);
+
+					}
+
+
+				// Output
+
+					return $item_output;
+
+			} // /wm_nav_menu_social_icons
+
+			add_filter( 'walker_nav_menu_start_el', 'wm_nav_menu_social_icons', 10, 4 );
+
+
+
+			/**
+			 * Display social links in Navigation Menu widget
+			 *
+			 * @since    2.5.0
+			 * @version  2.5.0
+			 *
+			 * @param  array  $nav_menu_args Array of parameters for `wp_nav_menu()` function.
+			 * @param  string $nav_menu      Menu slug assigned in the widget.
+			 * @param  array  $args          Widget parameters.
+			 */
+			function wm_social_widget( $nav_menu_args, $nav_menu, $args ) {
+
+				// Variables
+
+					$nav_menu_obj = wp_get_nav_menu_object( $nav_menu );
+					$locations    = get_nav_menu_locations();
+
+
+				// Requirements check
+
+					if (
+						! isset( $locations['social'] )
+						|| ! $locations['social']
+						|| absint( $locations['social'] ) !== absint( $nav_menu_obj->term_id )
+					) {
+						return $nav_menu_args;
+					}
+
+
+				// Processing
+
+					$nav_menu_args['container_class'] = 'social-links';
+					$nav_menu_args['menu_class']      = 'social-links-items';
+					$nav_menu_args['depth']           = 1;
+					$nav_menu_args['link_before']     = '<span class="screen-reader-text">';
+					$nav_menu_args['link_after']      = '</span>';
+					$nav_menu_args['items_wrap']      = '<ul id="%1$s" class="%2$s">%3$s</ul>';
+
+
+				// Output
+
+					return $nav_menu_args;
+
+			} // /wm_social_widget
+
+			add_filter( 'widget_nav_menu_args', 'wm_social_widget', 10, 3 );
 
 
 
@@ -2670,7 +2872,7 @@
 		 * Sticky post label
 		 *
 		 * @since    1.0
-		 * @version  2.0
+		 * @version  2.5.0
 		 */
 		if ( ! function_exists( 'wm_sticky_label' ) ) {
 			function wm_sticky_label() {
@@ -2678,7 +2880,7 @@
 				// Output
 
 					if ( is_sticky() ) {
-						echo '<div class="label-sticky" title="' . esc_attr__( 'This is sticky post', 'auberge' ) . '"><i class="genericon genericon-pinned"></i></div>';
+						echo '<div class="label-sticky" title="' . esc_attr__( 'This is sticky post', 'auberge' ) . '"><i class="genericons-neue genericons-neue-pinned"></i></div>';
 					}
 
 			}
@@ -2696,7 +2898,7 @@
 		 * If the post has more tag, display the content appropriately.
 		 *
 		 * @since    1.0
-		 * @version  2.0
+		 * @version  2.5.0
 		 *
 		 * @param  string $excerpt
 		 */
@@ -2723,12 +2925,6 @@
 						/**
 						 * Post has more tag
 						 */
-
-							// Required for <!--more--> tag to work
-
-								global $more;
-
-								$more = 0;
 
 							if ( has_excerpt() ) {
 								$excerpt = '<p class="entry-summary has-more-tag">' . get_the_excerpt() . '</p>';